style: clippy -D warnings back to zero after the wave (I4)
Six lints across the wave's new code: `then_some` over a lazy closure, two
useless `u64`/`f64` conversions, an unnecessary `usize` cast, two
`clone`-on-`Copy` calls, and `FRAC_1_SQRT_2` written out as a literal.

One of them was a real signal rather than a style note. `assert!(MAX_FRAME_MARKS
>= NEEDED)` was an assertion over two constants — clippy's "this assertion has a
constant value" is the same objection this tree makes as "a check that cannot
fail". It is a `const _: () = assert!(...)` now, which is strictly stronger (it
fails the BUILD rather than a test), and the runtime twin that cannot be folded
away already exists beside it: `the_report_names_every_pass_and_the_segments_tile_the_frame`
reads `EngineRenderer::pass_names` off the graph the renderer actually built.

+++ runtime/inf-player/tests/baked_building_geometry.rs
@@ -333,7 +333,7 @@ fn the_cost_of_the_city_as_cubes_and_as_baked_meshes() {
     );
     assert_eq!(
         vgeom_scene.vgeom_instances.len(),
-        (side * side) as usize,
+        side * side,
         "the vgeom side must draw one instance per building and no more"
     );
     let _ = BUILDINGS;
